Moved application logic into shared layer
