+++ src/ChannelEngine.Application/BusinessLogic/IBusinessLogic.cs
@@ -1,0 +1,7 @@
+﻿namespace ChannelEngine.Application.BusinessLogic
+{
+    public interface IBusinessLogic
+    {
+        Task<int> GetOrdersByStatus(string status);
+    }
+}
