+++ src/ChannelEngine.Application/ChannalEngineApi/Responses/OrderItemsResponse.cs
@@ -1,7 +1,6 @@
-﻿using ChannelEngine.External.Responses;
-using System.Text.Json.Serialization;
+﻿using System.Text.Json.Serialization;
 
-namespace ChannelEngine.ExternalApi.Responses
+namespace ChannelEngine.Application.ChannalEngineApi.Responses
 {
     public record OrderItemsResponse
     {
