+++ src/ChannelEngine.Application/Exceptions/OrderUriIsMissingException.cs
@@ -1,0 +1,8 @@
+﻿namespace ChannelEngine.Application.Exceptions
+{
+    public class OrderUriIsMissingException : ChannelEngineException
+    {
+        public OrderUriIsMissingException()
+            : base("Uri is missing exception.") { }
+    }
+}
