+++ src/ChannelEngine/ChannelEngine.csproj
@@ -30,4 +30,8 @@
     <ProjectReference Include="..\ChannelEngine.Application\ChannelEngine.Application.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="ChannalEngineApi\" />
+  </ItemGroup>
+
 </Project>
