--- src/ChannelEngine/Exceptions/OrderUriIsMissingException.cs
@@ -1,8 +1,0 @@
-﻿namespace ChannelEngine.Exceptions
-{
-    internal class OrderUriIsMissingException : ChannelEngineException
-    {
-        public OrderUriIsMissingException()
-            : base("Uri is missing exception.") { }
-    }
-}
